Move catalog fetch to custom hook
