+++ client/src/App.js
@@ -1,35 +1,20 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import Header from './components/Header'
 import Footer from './components/Footer.js'
 import { Route } from 'react-router-dom'
 import RootPage from './components/pages/RootPage'
 import DatasetDetailsPage from './components/pages/DatasetDetailsPage'
+import { useDatasetsCatalog } from './hooks/useDatasetsCatalog'
 
 function App() {
-  const [searchTerm, setSearchTerm] = useState()
-  const [datasetsCatalog, setDatasetsCatalog] = useState()
-  const [pageToFetch, setPageToFetch] = useState(1)
-  const [hasMoreDatasets, setHasMoreDatasets] = useState()
-
-  function increasePageToFetch() {
-    setPageToFetch(pageToFetch + 1)
-  }
-
-  function resetDatasetsCatalog() {
-    setDatasetsCatalog('')
-  }
-
-  useEffect(() => {
-    const pageQuery = 'page=' + pageToFetch
-    const searchTermQuery = searchTerm ? 'searchterm=' + searchTerm : ''
-    fetch('/datasets?' + pageQuery + '&' + searchTermQuery)
-      .then((response) => response.json())
-      .then((data) => {
-        setDatasetsCatalog(data.datasets)
-        setHasMoreDatasets(data.hasNextPage)
-      })
-      .catch((error) => console.log('error', error))
-  }, [searchTerm, pageToFetch])
+  const {
+    datasetsCatalog,
+    hasMoreDatasets,
+    resetDatasetsCatalog,
+    loadNextPage,
+    setSearchTerm,
+    searchTerm,
+  } = useDatasetsCatalog()
 
   return (
     <>
@@ -38,7 +23,7 @@ function App() {
         <RootPage
           datasetsCatalog={datasetsCatalog}
           setSearchTerm={setSearchTerm}
-          increasePageToFetch={increasePageToFetch}
+          loadNextPage={loadNextPage}
           hasMoreDatasets={hasMoreDatasets}
           resetDatasetsCatalog={resetDatasetsCatalog}
           searchTerm={searchTerm}
